refactor: Introduce the [EpubWriter] trait
The purpose of this trait is to abstract the medium onto which the
epub is generated. Previously, the use of the filesystem was
hard-coded. With this patch, we add an abstraction level which will
eventually allows us from using the same code to generate a zip
archive or an in-file-system directory.