feature: Generate the epub archive
Previously, the content of the epub was generated in a dedicated build
directory, and it was necessary to make the final step (generating the
archive) manually. With this patch, the archive is generated, but no
compression method is performed.

We add a new dependency to the `zip-rs' crate, since the latter
appears to be the go-to solution for generating zip archives.