chore: Abstract away the loading process of a Book project
At some point, we will want to create a website to write ogmarkup
document in a collaborative manner. To prepare that time, we abstract
away the loading process of a Book project with a dedicated trait.

In addition, the project is now divided into a library on the one
hand, and the celtchar executable on the other hand. The celtchar
executable uses its own implementation of the [Loader] trait to load a
Book project from the filesystem, as before. This means, currently,
the related functions are not provided by the celtchar library.