feature: List dependencies of a celtchar document
We provide an additional command `deps' that basically print one
dependency of —an ogmarkup document used to build a— celtchar
document. This way, `celtchar deps' can be used e.g., in a Makefile.

If a filename cannot be converted into a valid rust string, the
message <invalid utf8 filename> is print.